refactor: rename network.rs to network_state.rs
